feat: add custom file resolver
