+++ packages/certification-service/code/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@inditextech/apicertification",
-  "version": "1.1.0",
+  "version": "1.2.0",
   "description": "API scoring service",
   "author": "InditexTech Open Source Office",
   "license": "Apache-2.0",
